+++ py-rattler/rattler/solver/solver.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 import datetime
-from typing import List, Optional, Literal
+from typing import List, Optional, Literal, Sequence
 
-from rattler import Channel, Platform
+from rattler import Channel, Platform, VirtualPackage
 from rattler.match_spec.match_spec import MatchSpec
 
 from rattler.channel import ChannelPriority
@@ -18,17 +18,18 @@ SolveStrategy = Literal["highest", "lowest", "lowest-direct"]
 
 
 async def solve(
-    channels: List[Channel | str],
-    specs: List[MatchSpec | str],
+    channels: Sequence[Channel | str],
+    specs: Sequence[MatchSpec | str],
     gateway: Gateway = Gateway(),
-    platforms: Optional[List[Platform | PlatformLiteral]] = None,
-    locked_packages: Optional[List[RepoDataRecord]] = None,
-    pinned_packages: Optional[List[RepoDataRecord]] = None,
-    virtual_packages: Optional[List[GenericVirtualPackage]] = None,
+    platforms: Optional[Sequence[Platform | PlatformLiteral]] = None,
+    locked_packages: Optional[Sequence[RepoDataRecord]] = None,
+    pinned_packages: Optional[Sequence[RepoDataRecord]] = None,
+    virtual_packages: Optional[Sequence[GenericVirtualPackage | VirtualPackage]] = None,
     timeout: Optional[datetime.timedelta] = None,
     channel_priority: ChannelPriority = ChannelPriority.Strict,
     exclude_newer: Optional[datetime.datetime] = None,
     strategy: SolveStrategy = "highest",
+    constraints: Optional[List[MatchSpec | str]] = None,
 ) -> List[RepoDataRecord]:
     """
     Resolve the dependencies and return the `RepoDataRecord`s
@@ -69,6 +70,9 @@ async def solve(
             * `"lowest-direct"`: Select the lowest compatible version for all
               direct dependencies but the highest compatible version of transitive
               dependencies.
+        constraints: Additional constraints that should be satisfied by the solver.
+            Packages included in the `constraints` are not necessarily installed,
+            but they must be satisfied by the solution.
 
     Returns:
         Resolved list of `RepoDataRecord`s.
@@ -90,12 +94,23 @@ async def solve(
             gateway=gateway._gateway,
             locked_packages=[package._record for package in locked_packages or []],
             pinned_packages=[package._record for package in pinned_packages or []],
-            virtual_packages=[v_package._generic_virtual_package for v_package in virtual_packages or []],
+            virtual_packages=[
+                v_package.into_generic()._generic_virtual_package
+                if isinstance(v_package, VirtualPackage)
+                else v_package._generic_virtual_package
+                for v_package in virtual_packages or []
+            ],
             channel_priority=channel_priority.value,
             timeout=int(timeout / datetime.timedelta(microseconds=1)) if timeout else None,
             exclude_newer_timestamp_ms=int(exclude_newer.replace(tzinfo=datetime.timezone.utc).timestamp() * 1000)
             if exclude_newer
             else None,
             strategy=strategy,
+            constraints=[
+                constraint._match_spec if isinstance(constraint, MatchSpec) else PyMatchSpec(str(constraint), True)
+                for constraint in constraints
+            ]
+            if constraints is not None
+            else [],
         )
     ]